Fix: like btn not working on all pages
