+++ README.md
@@ -1,28 +1,44 @@
 # JavaScript Capstone Project
 
-In this project, we built a simple meals website were you can like your favorite meals and add comments.
+In this project, we built our own web application based on an external API. The web application has 2 user interfaces:
+
+- On homepage interface you can you can like your favorite meals;
+- On comments interface you can add comments;
 
 ## Built With
+
 - Java Script
 - HTML
 - CSS
+- Mealdb API
 
 ### Technologies used
+
 - Github
 - GitFlow
 - Visual Studio Code
 - webpack
 - npm
 - jest
+
 ## Deployment Link
 
-https://github.com/ShoiraTa/RecipeApp.git
+[Live demo](https://shoirata.github.io/RecipeApp/)
+![](demo.gif)
 
-## Live Version 
+## Getting Started
 
-## Website Demo
+To get a local copy up and running follow these simple example steps.
 
-## Authors
+- go to the directory that you want to clone the repository in.
+- right click and choose Git Bash here.
+- type "git clone https://github.com/ShoiraTa/RecipeApp.git".
+- run `npm install` to install all the dependencies
+- Run `npm start` to bundle the files in the src folder
+
+or
+
+- Go to the dist folder and open index.html
 
 👤 **Shoira Toshpulatova**
 
@@ -33,15 +49,19 @@ https://github.com/ShoiraTa/RecipeApp.git
 
 👤 **Daniel Shdeed**
 
-- GitHub: [https://github.com/Danieldotcomcoder)]
-- Twitter: [https://twitter.com/DannyDotcoder)]
-- LinkedIn: [https://www.linkedin.com/in/daniel-shdeed-832b03115/]
+- GitHub: [@githubhandle](https://github.com/Danieldotcomcoder))
+- LinkedIn: [LinkedIn](https://www.linkedin.com/in/daniel-shdeed-832b03115/)
+
+![Daniel's GitHub stats](https://github-readme-stats.vercel.app/api?username=danieldotcomcoder&count_private=true&theme=dark&show_icons=true)
+
 ## Acknowledgments
 
 - Microverse
+
 ## 📝 License
 
 This project is [MIT](./MIT) licensed.
 
 ## Show your support
+
 Give a ⭐️ if you like this project!
