+++ src/meal-api.js
@@ -1,66 +1,5 @@
-const getComments = async () => {
-  const submitBtn = document.getElementById('submit-comment');
-  const ul = document.getElementById('comments-ul');
-  const id = submitBtn.getAttribute('data');
-  const commentsCount = document.getElementById('comments-count');
-
-  const get = () => fetch(`https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/USJEEv7W2YAs453iqIPM/comments?item_id=${id}`)
-    .then((res) => res.json());
-
-  const mealComments = await get();
-
-  commentsCount.innerHTML = `${mealComments.length > 0 ? `${mealComments.length}` : '0'}`;
-
-  ul.innerHTML = '';
-  if (mealComments.length > 0) {
-    mealComments.forEach((elem) => {
-      const liComments = document.createElement('li');
-      liComments.innerHTML = `
-      <p><span class="bold">${elem.creation_date} ${elem.username}</span>: ${elem.comment}</p>
-      `;
-      ul.appendChild(liComments);
-    });
-  }
-};
-
-const postComment = () => {
-  const userNameInput = document.getElementById('input-name');
-  const alert = document.getElementById('alert');
-  const userComment = document.getElementById('comment');
-  const submitBtn = document.getElementById('submit-comment');
-
-  const post = (id, name, comment) => fetch('https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/USJEEv7W2YAs453iqIPM/comments',
-    {
-      method: 'POST',
-      headers: {
-        'Content-type': 'application/json; charset=UTF-8',
-      },
-      body: JSON.stringify(
-        {
-          item_id: id,
-          username: name,
-          comment,
-        },
-      ),
-    })
-    .then((res) => res.text());
-
-  getComments();
-  submitBtn.addEventListener('click', async (e) => {
-    e.preventDefault();
-
-    const id = submitBtn.getAttribute('data');
-    const name = userNameInput.value;
-    const comment = userComment.value;
-    alert.innerHTML = '';
-    if (name !== '' && comment !== '') {
-      userNameInput.value = '';
-      userComment.value = '';
-      await post(id, name, comment)
-        .then(() => getComments());
-    } else alert.innerHTML = 'Please insert your name and comment';
-  });
-};
+import { like, Displaylikes } from './interactions.js';
+import postComment from './comments-api.js';
 
 // Popup window //
 const popup = async () => {
@@ -179,10 +118,10 @@ const getmeals = async () => {
         </div>
         <div class="meal-body d-flex justify-between">
           <h4>${element.strMeal}</h4>
-          <button class="fav-btn" ><i class="fas fa-heart"></i></button>
+          <button class="fav-btn" ><i class="fas fa-heart" data="${element.idMeal}"></i></button>
         </div>
         <div>
-          <div class = "likes"><span class= "likes-qty"> </span> likes </div>
+          <div class = "likes"><span class= "likes-qty" data="${element.idMeal}">0</span> likes </div>
           <button class= "comments" data="${element.idMeal}">Comments</button>
         </div>
       </div>`;
@@ -228,7 +167,8 @@ const getmeals = async () => {
       getCategoryUrl(category);
       printCount(category);
       resetLinks();
-
+      setTimeout(() => like(), 2000);
+      Displaylikes();
       links[i].classList.add('active');
     });
   }
@@ -236,4 +176,5 @@ const getmeals = async () => {
   printCount('seafood');
 };
 
+Displaylikes();
 getmeals();
